Added some additional data
also added in gusto. I'm currently using Intuit Payroll (formerly paycycle) and although it used to be a good service, it leaves a lot to be desired (ie no repeat reminders... I missed an tax-due email once for 10 days and got a $2,500 fine from the IRS.) Orrick has some similar docs to the Goodwin-Procter docs (but I like the GP doc generator better). Don't forget to update NDA's to include the new whistleblower clauses in the Defend Trade Secrets Act..